feat: Implement API for the Settings page

+++ api/src/controllers/processes/seeding/index.ts
@@ -3,10 +3,9 @@ import { ChildProcess, spawn } from "child_process";
 import { Request, Response, NextFunction } from "express";
 
 import { NODE_PROCESS, PROCESSES } from "../../../config/processes";
-import * as processModel from "../../../model/processes";
-import * as trackModel from "../../../model/track";
+import * as processModel from "../../../models/processes";
+import * as trackModel from "../../../models/track";
 import { ProcessMessage } from "../../../types";
-import { HttpError } from "../../../utils/error";
 import { SSE } from "../../../middlewares/sse";
 
 export const seedingSSE = new SSE({
@@ -17,40 +16,39 @@ export const seedingSSE = new SSE({
 // We need to keep the process in memory in order to be able to kill it later
 let process: ChildProcess | null = null;
 
-async function onMessage(message: ProcessMessage) {
-  const processState = await processModel.update(message);
-  seedingSSE.send(processState, "seeding");
-  console.log("[message] Done. Database has been seeded");
-}
-
-async function onError(err: Error) {
-  console.log("[error] (child process - seeding)", err);
-  seedingSSE.send({ name: "seeding", status: "failure" }, "seeding");
-  await processModel.destroy("seeding");
-}
-
-function onJobClose() {
-  process = null;
-}
-
 export async function startSeeding(
   req: Request,
   res: Response,
   next: NextFunction,
 ) {
-  try {
-    console.log("Starting db seeding ...");
+  async function onMessage(message: ProcessMessage) {
+    console.log("[message] Done. Database has been seeded");
+    seedingSSE.send(await processModel.queries.update(message), "seeding");
+  }
 
+  async function onError(err: Error) {
+    console.log("[error] (child process - seeding)", err);
+    seedingSSE.send({ name: "seeding", status: "failure" }, "seeding");
+    await processModel.queries.destroy("seeding");
+  }
+
+  function onJobClose() {
+    process = null;
+  }
+
+  try {
     if (process && process.pid) {
-      res.status(500).json({ message: "Db seeding is already running ..." });
+      res
+        .status(400)
+        .json({ code: 400, message: "Db seeding is already running ..." });
     } else {
-      await processModel.destroy("seeding");
-      const processState = await processModel.create({
+      await processModel.queries.destroy("seeding");
+      const processState = await processModel.queries.create({
         name: "seeding",
         status: "pending",
       });
       res.status(202).json({ results: processState });
-      seedingSSE.send(processState);
+      seedingSSE.send(processState, "seeding");
 
       process = spawn(
         NODE_PROCESS,
@@ -72,7 +70,7 @@ export async function getSeedingStatusAsSSE(
   next: NextFunction,
 ) {
   try {
-    seedingSSE.send(await processModel.read("seeding"), "seeding");
+    seedingSSE.send(await processModel.queries.read("seeding"), "seeding");
   } catch (err) {
     next(err);
   }
@@ -84,19 +82,15 @@ export async function stopSeeding(
   next: NextFunction,
 ) {
   try {
-    if (!process || !process.pid) {
-      throw new HttpError({
-        code: 404,
-        message: `Nothing to stop — the seeding process is not running`,
-      });
-    } else {
+    if (process && process.pid) {
       process.kill("SIGTERM");
-
       process = null;
-      await processModel.destroy("seeding");
-      await trackModel.destroyAll();
-      res.status(200).end();
     }
+
+    await processModel.queries.destroy("seeding");
+    await trackModel.queries.destroyAll();
+    seedingSSE.send(null, "seeding");
+    res.status(204).end();
   } catch (err) {
     next(err);
   }
